+++ src/cpp/src/whisper/logit_processor.hpp
@@ -6,7 +6,6 @@
 #include <openvino/openvino.hpp>
 
 #include "openvino/genai/whisper_generation_config.hpp"
-#include "vector"
 
 namespace ov {
 namespace genai {
